GH-234 make sure all MarkupAnnotationComponent use enhanced contains detection for markup paths.
